More code cleanup.

+++ src/Engine.cpp
@@ -4,13 +4,9 @@
 
 #include "Engine.h"
 #include "graphics/TestObject.h"
-
 #include<SDL2/SDL.h>
 #include<GL/gl.h>
 #include <IL/il.h>
-#include <IL/ilu.h>
-#include <chrono>
-#include <iostream>
 
 using namespace MowerEngine;
 using namespace std;
@@ -49,8 +45,8 @@ void Engine::initSDL() {
     SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 2);
     SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 1);
 
-    window = SDL_CreateWindow("Mower Pre-Alpha", SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED, screenWidth,
-                              screenHeight, SDL_WINDOW_OPENGL | SDL_WINDOW_SHOWN);
+    window = SDL_CreateWindow("Mower Pre-Alpha", SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED, (int) screenWidth,
+                              (int) screenHeight, SDL_WINDOW_OPENGL | SDL_WINDOW_SHOWN);
     context = SDL_GL_CreateContext(window);
     SDL_GL_SetSwapInterval(1);
 }
@@ -131,12 +127,6 @@ void Engine::update() {
                 break;
         }
     }
-
-//    while (inputProcessor->hasProcessedMouseButtonInput()) {
-//        std::cout << (int) inputProcessor->getMouseButtonInput() << std::endl;
-//    }
-
-//    std::cout << inputProcessor->getMouseCoordinates().x << " " << inputProcessor->getMouseCoordinates().y << std::endl;
 
     inputProcessor->endProcessing();
     SDL_GL_SwapWindow(window);
